refactor: ReminderModal 네이밍 변경 -> AddReminder
