+++ frontend/src/components/@shared/Dropdown/index.tsx
@@ -1,6 +1,6 @@
 import { RefObject, useEffect } from "react";
 
-import useDropdown from "@src/components/Dropdown/@hooks/useDropdown";
+import useDropdown from "@src/components/@shared/Dropdown/@hooks/useDropdown";
 
 import useOuterClick from "@src/hooks/useOuterClick";
