+++ frontend/src/components/AddReminder/DateTimePicker/TimePicker/index.tsx
@@ -1,11 +1,11 @@
 import { ChangeEventHandler, RefObject } from "react";
 
+import Dropdown from "@src/components/@shared/Dropdown";
 import ReminderIconActive from "@src/components/@svgIcons/ReminderIconActive";
-import Dropdown from "@src/components/Dropdown";
-import { generateTimeOptions } from "@src/components/ReminderModal/@utils";
-import * as Styled from "@src/components/ReminderModal/DateTimePicker/DatePicker/style";
-import DateTimePickerOptions from "@src/components/ReminderModal/DateTimePicker/DateTimePickerOptions";
-import DateTimePickerToggle from "@src/components/ReminderModal/DateTimePicker/DateTimePickerToggle";
+import { generateTimeOptions } from "@src/components/AddReminder/@utils";
+import * as Styled from "@src/components/AddReminder/DateTimePicker/DatePicker/style";
+import DateTimePickerOptions from "@src/components/AddReminder/DateTimePicker/DateTimePickerOptions";
+import DateTimePickerToggle from "@src/components/AddReminder/DateTimePicker/DateTimePickerToggle";
 
 import { FlexColumn } from "@src/@styles/shared";
