+++ frontend/src/pages/Feed/index.tsx
@@ -2,6 +2,8 @@ import { Fragment } from "react";
 import { useLocation, useParams } from "react-router-dom";
 
 import InfiniteScroll from "@src/components/@shared/InfiniteScroll";
+import Modal from "@src/components/@shared/Modal";
+import AddReminder from "@src/components/AddReminder";
 import Calendar from "@src/components/Calendar";
 import DateDropdown from "@src/components/DateDropdown";
 import EmptyStatus from "@src/components/EmptyStatus";
@@ -9,9 +11,7 @@ import MessageCard from "@src/components/MessageCard";
 import BookmarkButton from "@src/components/MessageCard/MessageIconButtons/BookmarkButton";
 import ReminderButton from "@src/components/MessageCard/MessageIconButtons/ReminderButton";
 import MessagesLoadingStatus from "@src/components/MessageCard/MessagesLoadingStatus";
-import ReminderModal from "@src/components/ReminderModal";
 import SearchForm from "@src/components/SearchForm";
-import Modal from "@src/components/@shared/Modal";
 
 import useGetInfiniteMessages from "@src/hooks/query/useGetInfiniteMessages";
 import useMutateBookmark from "@src/hooks/query/useMutateBookmark";
@@ -144,7 +144,7 @@ function Feed() {
           handleCloseReminderModal();
         }}
       >
-        <ReminderModal
+        <AddReminder
           messageId={reminderTarget.id}
           remindDate={reminderTarget.remindDate ?? ""}
           handleCloseReminderModal={() => {
