+++ frontend/src/pages/SearchResult/index.tsx
@@ -2,11 +2,11 @@ import * as Styled from "@src/pages/Feed/style";
 
 import InfiniteScroll from "@src/components/@shared/InfiniteScroll";
 import Modal from "@src/components/@shared/Modal";
+import AddReminder from "@src/components/AddReminder";
 import MessageCard from "@src/components/MessageCard";
 import BookmarkButton from "@src/components/MessageCard/MessageIconButtons/BookmarkButton";
 import ReminderButton from "@src/components/MessageCard/MessageIconButtons/ReminderButton";
 import MessagesLoadingStatus from "@src/components/MessageCard/MessagesLoadingStatus";
-import ReminderModal from "@src/components/ReminderModal";
 import SearchForm from "@src/components/SearchForm";
 
 import useGetInfiniteMessages from "@src/hooks/query/useGetInfiniteMessages";
@@ -120,7 +120,7 @@ function SearchResult() {
           handleCloseReminderModal();
         }}
       >
-        <ReminderModal
+        <AddReminder
           messageId={reminderTarget.id}
           remindDate={reminderTarget.remindDate ?? ""}
           handleCloseReminderModal={() => {
